Fix wind warping divisor

+++ projects/dunes/dunes/core/simulator.hpp
@@ -43,7 +43,7 @@ namespace dunes
 		void setVenturiStrength(const float t_venturiStrength);
 		void setWindWarpingMode(const WindWarpingMode t_windWarpingMode);
 		void setWindWarpingCount(const int t_windWarpingCount);
-		void setWindWarpingDivisor(const int t_windWarpingDivisor);
+		void setWindWarpingDivisor(const float t_windWarpingDivisor);
 		void setWindWarpingRadius(const int t_index, const float t_windWarpingRadius);
 		void setWindWarpingStrength(const int t_index, const float t_windWarpingStrength);
 		void setWindWarpingGradientStrength(const int t_index, const float t_windWarpingGradientStrength);
